requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
mysql-connector-python==8.0.28
flwr>=1.0, <2.0
flwr-datasets[vision]>=0.0.2, <1.0.0
numpy==1.23.2
pandas==2.0.0
opencv-python==4.5.3.56
scikit-learn==0.24.2
